backend/requirements.txt
fastapi
uvicorn
python-dotenv
faiss-cpu
google-genai
pydantic
numpy